novedades clase 23 sept

+++ ciclo3backGrupo13/src/main/java/co/edu/unbusque/ciclo3backGrupo13/dao/UsuariosDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import co.edu.unbusque.ciclo3backGrupo13.model.Usuarios;
 
-public interface UsuariosDAO extends JpaRepository<Usuarios, Integer> {
+public interface UsuariosDAO extends JpaRepository<Usuarios, Long> {
 
 }
