+++ ciclo3backGrupo13/src/main/java/co/edu/unbusque/ciclo3backGrupo13/dao/productosDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import co.edu.unbusque.ciclo3backGrupo13.model.productos;
 
-public interface productosDAO extends JpaRepository<productos, Integer>{
+public interface productosDAO extends JpaRepository<productos, Long>{
 
 }
